Make *file() be defensive against a concurrent free since they use a shared fid

+++ fidclnt/fidclnt.go
@@ -171,7 +171,11 @@ func (fidc *FidClnt) Remove(fid np.Tfid) *np.Err {
 }
 
 func (fidc *FidClnt) RemoveFile(fid np.Tfid, wnames []string, resolve bool) *np.Err {
-	f := fidc.ft.Lookup(fidc.fids.lookup(fid).Path().AppendPath(wnames))
+	ch := fidc.fids.lookup(fid)
+	if ch == nil {
+		return np.MkErr(np.TErrUnreachable, "getfile")
+	}
+	f := fidc.ft.Lookup(ch.Path().AppendPath(wnames))
 	return fidc.fids.lookup(fid).pc.RemoveFile(fid, wnames, resolve, f)
 }
 
@@ -211,7 +215,11 @@ func (fidc *FidClnt) WriteV(fid np.Tfid, off np.Toffset, data []byte, v np.TQver
 }
 
 func (fidc *FidClnt) GetFile(fid np.Tfid, path []string, mode np.Tmode, off np.Toffset, cnt np.Tsize, resolve bool) ([]byte, *np.Err) {
-	f := fidc.ft.Lookup(fidc.fids.lookup(fid).Path().AppendPath(path))
+	ch := fidc.fids.lookup(fid)
+	if ch == nil {
+		return nil, np.MkErr(np.TErrUnreachable, "getfile")
+	}
+	f := fidc.ft.Lookup(ch.Path().AppendPath(path))
 	reply, err := fidc.fids.lookup(fid).pc.GetFile(fid, path, mode, off, cnt, resolve, f)
 	if err != nil {
 		return nil, err
@@ -220,7 +228,11 @@ func (fidc *FidClnt) GetFile(fid np.Tfid, path []string, mode np.Tmode, off np.T
 }
 
 func (fidc *FidClnt) SetFile(fid np.Tfid, path []string, mode np.Tmode, off np.Toffset, data []byte, resolve bool) (np.Tsize, *np.Err) {
-	f := fidc.ft.Lookup(fidc.fids.lookup(fid).Path().AppendPath(path))
+	ch := fidc.fids.lookup(fid)
+	if ch == nil {
+		return 0, np.MkErr(np.TErrUnreachable, "getfile")
+	}
+	f := fidc.ft.Lookup(ch.Path().AppendPath(path))
 	reply, err := fidc.fids.lookup(fid).pc.SetFile(fid, path, mode, off, resolve, f, data)
 	if err != nil {
 		return 0, err
@@ -229,7 +241,11 @@ func (fidc *FidClnt) SetFile(fid np.Tfid, path []string, mode np.Tmode, off np.T
 }
 
 func (fidc *FidClnt) PutFile(fid np.Tfid, path []string, mode np.Tmode, perm np.Tperm, off np.Toffset, data []byte) (np.Tsize, *np.Err) {
-	f := fidc.ft.Lookup(fidc.fids.lookup(fid).Path().AppendPath(path))
+	ch := fidc.fids.lookup(fid)
+	if ch == nil {
+		return 0, np.MkErr(np.TErrUnreachable, "getfile")
+	}
+	f := fidc.ft.Lookup(ch.Path().AppendPath(path))
 	reply, err := fidc.fids.lookup(fid).pc.PutFile(fid, path, mode, perm, off, f, data)
 	if err != nil {
 		return 0, err
